Pass-6 audit remediation (job 745: 0C/0H/4M/12L/15I)
- M3: convert() no longer bricks if one prize vault is QUOTRON-blacklisted
  (each leg via low-level send; split the full balance so a stuck slice auto-retries)
- M4: setPotCap re-pegs bounded to +/-25% per cooldown (no one-tx pot manipulation)
- L1: LP gate now covers REMOVE as well as add; hook re-mined 0x1844 -> 0x1A44
- L2: lockEngines rejects incomplete/mismatched binding sets
- L3: setAdapters frozen by lockRouting (no post-lock redirect)
- L4: deploy asserts HolderDraw/Raffle genesis match
- L5: claimBatch survives a single reverting payout; the rest still pay
- L7: int128-min-safe negation in QpullWethAdapter
- L9: BlsDrandOracle also probes MAP_FP_TO_G1 at construction
- SECURITY.md section 13 maps every finding (fixed / accepted / refuted)

+++ src/adapters/QpullWethAdapter.sol
@@ -144,7 +144,9 @@ contract QpullWethAdapter is ISwapAdapter, IUnlockCallback, Ownable2Step, Reentr
         // amountIn. On a partial fill (pool liquidity exhausted at the price-limit sentinel) V4 takes less than
         // amountIn; settling the nominal would leave a positive delta and revert unlock() with CurrencyNotSettled.
         int128 in128 = zeroForOne ? delta.amount0() : delta.amount1(); // negative = QPULL owed to the pool
-        uint256 consumed = uint256(uint128(-in128));
+        // audit L7 (job-745): promote to int256 BEFORE negating so type(int128).min can't overflow (matches
+        // the hook's own _abs idiom). Numerically identical for all reachable fills.
+        uint256 consumed = uint256(-int256(in128));
 
         // Pay QPULL in: sync, transfer the consumed amount, settle.
         poolManager.sync(Currency.wrap(qpull));
